correction user choice backend

+++ backend/app/counter/api/history_views.py
@@ -6,12 +6,14 @@ from decimal import Decimal
 from  app.counter.models import Counter
 from  app.counter.api.serializers import CounterSerializer
 
+import logging
+
 @api_view(['GET'])
 def api_get_all_history(request):
 	try:
 		all_history = Counter.objects.all()
 	except Counter.DoesNotExist:
-			return Response(0)
+		all_history = []
 
 	serializer = CounterSerializer(all_history, many=True)
 	return Response(data=serializer.data, status=status.HTTP_200_OK)
@@ -36,14 +38,18 @@ def api_delete_history_counter(request):
 
 	try:
 		historic_counter = Counter.objects.all()
-	except:
-		historic_counter = 0
+	except Exception as e:
+		logging.warning(e)
+		return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
-	operation_success = historic_counter.delete()
+	operation_success = False
+	if len(historic_counter) > 0:
+		operation_success = historic_counter.delete()
+
 	data = {}
 	if operation_success:
 		data['success'] = "delete all successful"
 	else:
 		data['success'] = "delete all failed"
 	
-	return Response(data=data)
+	return Response(data=data, status=status.HTTP_200_OK)
